[time]cast XCLAMP divisor arg to uint64_t in clockevent/clocksource drivers to fix compare-distinct-pointer-types warnings

+++ projects/arm32-baremetal-rv1103/driver/ce-armv7-timer.c
@@ -138,7 +138,7 @@ static struct device_t * ce_armv7_timer_probe(struct driver_t * drv, struct dtno
 
 	pdat->irq = irq;
 
-	clockevent_calc_mult_shift(ce, (uint32_t)rate, (uint32_t)XCLAMP(0x00000000ffffffffULL / (uint64_t)rate, (uint64_t)1, (uint64_t)600));
+	clockevent_calc_mult_shift(ce, (uint32_t)rate, (uint32_t)XCLAMP((uint64_t)(0x00000000ffffffffULL / (uint64_t)rate), (uint64_t)1, (uint64_t)600));
 	ce->name = alloc_device_name(dt_read_name(n), dt_read_id(n));
 	ce->min_delta_ns = clockevent_delta2ns(ce, 0x1);
 	ce->max_delta_ns = clockevent_delta2ns(ce, 0x00000000ffffffffULL);
